requirements.txt
flask
flask_sqlalchemy
sqlalchemy
sentence-transformers
scikit-learn
python-dotenv
PyPDF2
